+++ NEWS.md
@@ -6,6 +6,7 @@
 * `sg_get_cluster` added - helper to get nodes color by cluster.
 * `sg_button` added - Add buttons to trigger events in static documents.
 * `sg_export` and `sg_export_p` added - To export the graph.
+* `sg_progress` added - show dynamic text.
 
 # sigmajs 0.0.2